requirements.txt
torch>=2.0.0
transformers>=4.40.0
qwen-vl-utils>=0.0.8
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
PyYAML>=6.0
tqdm>=4.65.0
